pyproject.toml
[tool.poetry]
name = "udacity-project-3"
version = "0.1.0"
description = ""
authors = ["None"]

[tool.poetry.dependencies]
python = "^3.9"
psycopg2-binary = "^2.9.1"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.7"
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
